Pull render loop out of Renderer into Server.
This allows Renderer to only use platform-agnostic piston primitives.
